+++ scripts/targets/rauc/installer/home-skel/distr/install_hasplm.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+echo "installing hasplm..."
+bash /home/inauto/distr/install_tmp/hasplm/postinst
+
+echo "Installing hasplm complete"