+++ scripts/targets/rauc/installer/home-skel/distr/install_idmvs.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+echo "installing MVS..."
+bash /home/inauto/distr/install_tmp/MVS/setup.sh
+
+echo "installing IDMVS..."
+sh /home/inauto/distr/install_tmp/IDMVS/setup.sh
+
+echo "Installing addons complete"